Split ContractDeployer to its own file
I am trying to mitigate #714 now.
